ISSUE #1833 - Move all drawing handlers to separate components
