+++ frontend/routes/components/screenshotDialog/components/drawingHandler/drawingHandler.helpers.ts
